refactor: remove Generative UI components and documentation, update package dependencies
