--- examples/genui-live-proxy/.env.example
@@ -1,11 +1,0 @@
-OPENAI_API_KEY=sk-your-key-here
-OPENAI_BASE_URL=https://api.openai.com/v1
-OPENAI_MODEL=gpt-4o-mini
-PORT=8787
-
-# Corporate / internal upstreams: if you see `SELF_SIGNED_CERT_IN_CHAIN`,
-# trust the CA bundle here (preferred):
-# NODE_EXTRA_CA_CERTS=/absolute/path/to/corp-ca.pem
-#
-# Or, for a quick dev-only bypass (insecure; localhost only):
-# NODE_TLS_REJECT_UNAUTHORIZED=0
