--- src/genui/__stories__/Docs.mdx
@@ -1,79 +1,0 @@
-import {Meta, Canvas} from '@storybook/addon-docs';
-
-import * as Stories from './GenUIOverview.stories';
-
-<Meta of={Stories} />
-
-# Generative UI
-
-AIKit's GenUI layer lets a model emit structured `tool-call` parts that the chat
-maps to interactive React components, then surfaces a typed round-trip back to
-the model.
-
-It is **opt-in**: omit `genUIRegistry` and AIKit behaves exactly as before.
-
-## Two registries
-
-- `MessageRendererRegistry` (existing) is keyed by `content.type` — it owns how
-  each part type is rendered (`text`, `thinking`, `tool`, `tool-call`,
-  `tool-result`, plus any custom types you add).
-- `GenUIToolRegistry` (new) is keyed by **tool name** — it holds the schema,
-  component, and optional loading/error slots for each tool the model may call.
-
-When `genUIRegistry` is provided, the default `tool-call` renderer is installed
-into the `MessageRendererRegistry`; it looks each `tool-call` part up by
-`toolName` in the GenUI registry and dispatches to the matching component.
-
-## Quickstart
-
-```tsx
-import {z} from 'zod';
-import {createGenUIToolRegistry, registerGenUITool, ChatContainer} from '@gravity-ui/aikit';
-
-const registry = createGenUIToolRegistry();
-
-registerGenUITool(registry, {
-    name: 'weather.show',
-    description: 'Render a weather card for a city.',
-    schema: z.object({city: z.string(), units: z.enum(['c', 'f']).optional()}),
-    component: ({args, submitResult}) => (
-        <button onClick={() => submitResult({acknowledged: true})}>
-            {args.city} ({args.units ?? 'c'})
-        </button>
-    ),
-});
-
-<ChatContainer
-    /* ...usual props... */
-    messageListConfig={{
-        genUIRegistry: registry,
-        onToolResult: (event) => {
-            // append `event.part` to your conversation history
-        },
-    }}
-/>;
-```
-
-JSON Schema works the same way — pass a `JSONSchema7` instead of a Zod schema.
-
-## Lifecycle
-
-`tool-call.data.status` drives the renderer:
-
-- `input-streaming` — render `tool.loading ?? <DefaultLoadingSkeleton />`
-- `input-available` — validate args against the tool's schema, then mount the
-  component
-- `output-available` — same, but `previousResult` is rehydrated from a sibling
-  `tool-result` part with the same `toolCallId`
-- `output-error` — render `tool.error ?? <DefaultErrorSlot />`
-
-Every component is wrapped in a per-part error boundary; render crashes don't
-take down the sibling parts.
-
-## Lifecycle examples
-
-<Canvas of={Stories.Lifecycle} />
-
-## Round-trip and rehydration
-
-<Canvas of={Stories.RoundTrip} />
