--- src/genui/renderers/DefaultErrorSlot.scss
@@ -1,23 +1,0 @@
-.g-aikit-genui-error-slot {
-    display: flex;
-    flex-direction: column;
-    gap: var(--g-spacing-1, 4px);
-    padding: var(--g-spacing-2, 8px) var(--g-spacing-3, 12px);
-    border: 1px solid var(--g-color-line-danger, rgba(229, 57, 53, 0.4));
-    background: var(--g-color-base-danger-light, rgba(229, 57, 53, 0.08));
-    border-radius: var(--g-border-radius-m, 6px);
-    color: var(--g-color-text-danger, #d32f2f);
-    font-size: var(--g-text-body-1-font-size, 13px);
-
-    &__title {
-        font-weight: 600;
-        font-family: var(--g-font-family-monospace, monospace);
-        font-size: 12px;
-        text-transform: uppercase;
-        letter-spacing: 0.04em;
-    }
-
-    &__message {
-        color: var(--g-color-text-primary, inherit);
-    }
-}
