--- src/genui/renderers/DefaultLoadingSkeleton.scss
@@ -1,14 +1,0 @@
-.g-aikit-genui-loading-skeleton {
-    display: inline-flex;
-    align-items: center;
-    gap: var(--g-spacing-2, 8px);
-    padding: var(--g-spacing-2, 8px) var(--g-spacing-3, 12px);
-    border: 1px dashed var(--g-color-line-generic, rgba(0, 0, 0, 0.15));
-    border-radius: var(--g-border-radius-m, 6px);
-    color: var(--g-color-text-secondary, rgba(0, 0, 0, 0.6));
-    font-size: var(--g-text-body-1-font-size, 13px);
-
-    &__label {
-        font-family: var(--g-font-family-monospace, monospace);
-    }
-}
