--- src/genui/renderers/UnknownToolFallback.scss
@@ -1,24 +1,0 @@
-.g-aikit-genui-unknown-tool {
-    display: flex;
-    flex-direction: column;
-    gap: var(--g-spacing-1, 4px);
-    padding: var(--g-spacing-2, 8px) var(--g-spacing-3, 12px);
-    border: 1px dashed var(--g-color-line-warning, rgba(255, 152, 0, 0.5));
-    background: var(--g-color-base-warning-light, rgba(255, 152, 0, 0.08));
-    border-radius: var(--g-border-radius-m, 6px);
-    color: var(--g-color-text-warning, #ed6c02);
-    font-size: var(--g-text-body-1-font-size, 13px);
-
-    &__title {
-        font-weight: 600;
-        font-family: var(--g-font-family-monospace, monospace);
-    }
-
-    &__args {
-        margin: 0;
-        white-space: pre-wrap;
-        word-break: break-word;
-        font-size: 12px;
-        color: var(--g-color-text-primary, inherit);
-    }
-}
