+++ src/index.ts
@@ -23,9 +23,6 @@ export * from './components/pages';
 // === Adapters ===
 export * from './adapters';
 
-// === Generative UI ===
-export * from './genui';
-
 // === Hooks ===
 export * from './hooks';
